% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_travis}
\alias{auth_travis}
\title{Authorize Travis-CI API}
\usage{
auth_travis(username, password, setenv = TRUE, clean = !missing(username),
  force = FALSE, gh_token = c(Sys.getenv("GITHUB_TOKEN"),
  Sys.getenv("GITHUB_PAT")), base = c("https://api.travis-ci.org",
  "https://api.travis-ci.com"))
}
\arguments{
\item{username}{A character string containing a GitHub username.}

\item{password}{A character string containing a GitHub password for account \code{username}.}

\item{setenv}{A logical indicating whether to set the \env{TRAVIS_CI_TOKEN} environment variable. If \code{FALSE}, this will have to be passed to all other functions.}

\item{clean}{A logical specifying whether to delete the temporarily generated GitHub personal access token from the GitHub account. Deafult is \code{TRUE} if \code{username} is specified. This is ignored if passing a GitHub token directly.}

\item{force}{A logical indicating whether to force the operation and override an already stored value in \env{TRAVIS_CI_TOKEN}.}

\item{gh_token}{An optional character string  specifying a GitHub personal access token. This is ignored if \code{username} is supplied.}

\item{base}{A character string specifying the base URL for the API. By default this is \url{https://api.travis-ci.org}, but may need to be changed if using a Travis Pro or enterprise account.}
}
\value{
A character string containing the Travis-CI API token, invisibly.
}
\description{
Initialize a Travis-CI API token
}
\details{
Using the Travis-CI API requires an API token (see \href{http://docs.travis-ci.com/api/#authentication}{API documentation} for details). This function implements a sort of handshake that uses a GitHub personal access token to generate a Travis token (to store in an environment variable, \env{TRAVIS_CI_TOKEN}) or, alternatively, uses a GitHub username and password pair to generate such a token, conduct the handshake, and then cleanup the generated, temporary token. The easiest way to use the function is to set the \env{GITHUB_TOKEN} and \env{GITHUB_PAT} environment variables, then simply call \code{auth_travis()} with no arguments, but see examples.
}
\examples{
\dontrun{
# authenticate using GitHub login credentials
auth_travis("username", "password")

# authenticate using a GitHub personal access token
auth_travis(gh_token = "example token")

# set GITHUB_TOKEN environment variable to authenticate
Sys.setenv("GITHUB_TOKEN" = "example token")
auth_travis()
}

}
\seealso{
\code{\link{travisHTTP}}
}
