% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/builds.R
\name{cancel_build}
\alias{cancel_build}
\alias{restart_build}
\alias{restart_last_build}
\title{Cancel and Restart Builds}
\usage{
cancel_build(build, ...)

restart_build(build, ...)

restart_last_build(repo, ...)
}
\arguments{
\item{build}{A numeric value specifying a build number.}

\item{...}{Additional arguments passed to \code{\link{travisHTTP}}.}
}
\value{
For \code{cancel_build} and \code{restart_build}, a logical that is \code{TRUE} if the operation succeeded. For \code{restart_last_build}, the build number is stored in the \code{build_id} attributes.
}
\description{
Cancel and restart Travis-CI builds
}
\details{
\code{cancel_build} will cancel a given build. \code{restart_build} will restart a cancelled build.
}
\examples{
\dontrun{
# authenticate based on Sys.setenv("GITHUB_TOKEN" = "sometoken")
auth_travis()

# get all recent builds for this repo
g <- get_builds(repo = "cloudyr/travisci")

# cancel most recent build
cancel_build(g[[1]]$id)

# restart that build
restart_build(g[[1]]$id)
}
}
\seealso{
\code{\link{get_builds}}
}

