% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/account.R
\name{get_accounts}
\alias{get_accounts}
\title{Get Travis Accounts}
\usage{
get_accounts(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{travisHTTP}}.}
}
\value{
A list.
}
\description{
Retrieve GitHub accounts linked to the authenticated Travis user.
}
\details{
This is probably the closest thing to a \dQuote{Hello World!} on the API. It provides information about what accounts have been linked between GitHub and Travis, such as organization accounts that the user is a member of.
}
\examples{
\dontrun{
# authenticate based on Sys.setenv("GITHUB_TOKEN" = "sometoken")
auth_travis()

# get travis accounts
get_accounts()
}
}
\seealso{
\code{\link{get_users}}, \code{\link{get_permissions}}
}

