% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{get_annotations}
\alias{get_annotations}
\alias{create_annotation}
\title{Job Annotations}
\usage{
get_annotations(job, ...)

create_annotation(job, body = list(), ...)
}
\arguments{
\item{job}{A numeric job ID, or an object of class \dQuote{travis_job}.}

\item{...}{Additional arguments passed to \code{\link{travisHTTP}}.}

\item{body}{A list of arguments specifying the annotation. See \href{http://docs.travis-ci.com/api/#annotations}{API documentation} for details.}
}
\value{
A list.
}
\description{
Retrieve and create job Travis CI annotations
}
\details{
\code{get_annotations} retrieves any annotations attached to a Travis CI job. \code{create_annotation} creates a new annotation for a job. Note: As of October, 2015 support for annotations is considered \dQuote{experimental}.
}
\examples{
\dontrun{
# authenticate based on Sys.setenv("GITHUB_TOKEN" = "sometoken")
auth_travis()

# get annotations for a job
get_annotations(12345)

# create a new annotation
create_annotations(12345, body = list(description = "job description", status = "passed"))
}
}
\seealso{
\code{\link{get_job}}

\code{\link{get_job}}, \code{\link{get_logs}}
}
