% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{get_broadcasts}
\alias{get_broadcasts}
\title{Get Broadcasts}
\usage{
get_broadcasts(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{travisHTTP}}.}
}
\value{
A list.
}
\description{
Retrieve Travis-CI broadcasts
}
\details{
This is not particularly useful from an R perspective, but it will retrieve \dQuote{broadcasts} or news from Travis-CI.
}
\examples{
\dontrun{
# authenticate based on Sys.setenv("GITHUB_TOKEN" = "sometoken")
auth_travis()

# get broadcasts
get_broadcasts()
}
}
