% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/builds.R
\name{get_builds}
\alias{get_builds}
\title{Get Builds}
\usage{
get_builds(repo, build, ...)
}
\arguments{
\item{repo}{Optionally, a numeric repository ID (such as returned by this function), a character string specifying a GitHub repository \dQuote{slug} (e.g., \samp{ghusername/ghreponame}), or an object of class \dQuote{travis_repo}.}

\item{build}{Optionally, a numeric value specifying a build number, or an object of class \dQuote{travis_build}.}

\item{...}{Additional arguments passed to \code{\link{travisHTTP}}.}
}
\value{
A list.
}
\description{
Retrieve Travis Builds
}
\details{
This can retrieve a list of recent builds (across all repos), recent builds for a specific repo (if \code{repo} is specified), or information about a specific build if \code{build} is (or both \code{repo} and \code{build} are) specified.
}
\examples{
\dontrun{
# authenticate based on Sys.setenv("GITHUB_TOKEN" = "sometoken")
auth_travis()

# get all recent builds
g <- get_builds()

# get builds for a specific repo
get_builds(repo = "cloudyr/travisci")
get_builds(repo = g[[1]]$repository_id)

# get a specific build
get_builds(build = g[[1]]$id)
}
}
\seealso{
\code{\link{cancel_build}}, \code{\link{restart_build}}

\code{\link{cancel_build}}, \code{\link{get_repo}}, \code{\link{get_branch}}
}

