% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caches.R
\name{get_caches}
\alias{delete_caches}
\alias{get_caches}
\title{Get and Delete Cache}
\usage{
get_caches(repo, ...)

delete_caches(repo, ...)
}
\arguments{
\item{repo}{A numeric repository ID (such as returned by this function), a character string specifying a GitHub repository \dQuote{slug} (e.g., \samp{ghusername/ghreponame}), or an object of class \dQuote{travis_repo}.}

\item{...}{Additional arguments passed to \code{\link{travisHTTP}}.}
}
\value{
A list.
}
\description{
Retrieve the Travis-CI cache, or delete it.
}
\details{
\code{get_caches} retrieves caches for a repository. \code{delete_caches} deletes all caches for a repository.
}
\examples{
\dontrun{
# authenticate based on Sys.setenv("GITHUB_TOKEN" = "sometoken")
auth_travis()

# get caches for a repo
get_caches("cloudyr/travisci")

# delete caches
delete_caches("cloudyr/travisci")
}
}

