% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/environment.R
\name{get_encryption_key}
\alias{get_encryption_key}
\alias{set_encryption_key}
\title{Manage encrypted environment variables}
\usage{
get_encryption_key(repo, ...)

set_encryption_key(repo, ...)
}
\arguments{
\item{repo}{A character string specifying a repo slug (i.e., \samp{cloudyr/travisci}), a numeric Travis-CI repository ID, or an object of class \dQuote{travis_repo}. If a slug is used, it will be implicitly converted to a repository ID in some cases where only the latter is accepted.}

\item{...}{Additional arguments passed to \code{\link{travisHTTP}}.}
}
\value{
A list.
}
\description{
Retrieve and refresh the public key used for encryption.
}
\details{
These functions retrieve and refresh the public key used for encrypting secure environment variables for the Travis-CI build environment. This is probably more easily handled using the travis command-line tools, but it is also an implemented feature of the API and the functionality may be expanded in the future. See href{http://docs.travis-ci.com/user/encryption-keys/}{the API documentation} for full details. \code{get_encryption_key} retrieves the current, repository-specific public key. \code{set_encryption_key} resets the key and returns its value.
}
\examples{
\dontrun{
# authenticate based on Sys.setenv("GITHUB_TOKEN" = "sometoken")
auth_travis()

# get current public key
get_encryption_key("cloudyr/travisci")

# reset public key and retrieve the new value
set_encryption_key("cloudyr/travisci")
}
}
\seealso{
\code{\link{get_env_vars}}
}

