% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hooks.R
\name{get_hooks}
\alias{disable_hook}
\alias{enable_hook}
\alias{get_hooks}
\title{Get Hooks}
\usage{
get_hooks(...)

enable_hook(hook, ...)

disable_hook(hook, ...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{travisHTTP}}.}

\item{hook}{A numeric hook ID, or an object of class \dQuote{travis_hook}.}
}
\value{
A list.
}
\description{
Retrieve a list of Travis-CI web hooks.
}
\details{
\code{get_hooks} retrieves a list of hooks. \code{enable_hook} and \code{disable_hook} enable and disable hooks, respectively.
}
\examples{
\dontrun{
# authenticate based on Sys.setenv("GITHUB_TOKEN" = "sometoken")
auth_travis()

# get list of hooks
get_hooks()

# disable a hook
disable_hook("12345")

# enable a hook
enable_hook("1235")
}
}

