% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jobs.R
\name{get_job}
\alias{cancel_job}
\alias{get_job}
\alias{restart_job}
\title{Get Job}
\usage{
get_job(job, ...)

cancel_job(job, ...)

restart_job(job, ...)
}
\arguments{
\item{job}{A numeric job ID.}

\item{...}{Additional arguments passed to \code{\link{travisHTTP}}.}
}
\value{
A list.
}
\description{
Retrieve a Travis-CI job
}
\details{
\code{get_job} retrieves a list of details about a given job. \code{cancel_job} cancels and \code{restart_job} restarts a given job.
}
\examples{
\dontrun{
# authenticate based on Sys.setenv("GITHUB_TOKEN" = "sometoken")
auth_travis()

# get a job
get_job("12345")

# cancel a job
cancel_job("12345")

# restart a job
restart_job("1235")
}
}
\seealso{
\code{\link{get_annotations}}, \code{\link{get_logs}}
}

