% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repo.R
\name{get_repo}
\alias{get_repo}
\title{Get Repo}
\usage{
get_repo(repo, ...)
}
\arguments{
\item{repo}{A numeric repository ID (such as returned by this function) or a character string specifying a GitHub repository \dQuote{slug} (e.g., \samp{ghusername/ghreponame}).}

\item{...}{Additional arguments passed to \code{\link{travisHTTP}}.}
}
\value{
A list.
}
\description{
Retrieve a Travis-CI repository
}
\details{
This retrieves a list of details about a given repository.
}
\examples{
\dontrun{
# get a repo
get_repo(repo = "cloudyr/travisci")

# get a specific branch
get_branch(repo = "cloudyr/travisci", "master")

# get recent builds for that repo
get_builds(repo = "cloudyr/travisci")
}
}
\seealso{
\code{\link{get_builds}}

\code{\link{get_builds}}, \code{\link{get_branch}}
}

