% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{get_requests}
\alias{get_requests}
\title{GitHub Requests}
\usage{
get_requests(repo, request, ...)
}
\arguments{
\item{repo}{Optionally, a numeric repository ID (such as returned by this function), a character string specifying a GitHub repository \dQuote{slug} (e.g., \samp{ghusername/ghreponame}), or an object of class \dQuote{travis_repo}. Must specify either \code{repo} or \code{request}.}

\item{request}{Optionally, a numeric request ID. Must specify either \code{repo} or \code{request}.}

\item{...}{Additional arguments passed to \code{\link{travisHTTP}}.}
}
\value{
A list.
}
\description{
Examine requests from GitHub to Travis CI
}
\details{
This can examine requests made by GitHub to initiate builds on Travis-CI, perhaps when a push fails to execute a build.
}
\examples{
\dontrun{
# authenticate based on Sys.setenv("GITHUB_TOKEN" = "sometoken")
auth_travis()

# get a list of requests
get_requests()

# get a specific request
get_requests(12345)
}
}
\seealso{
\code{\link{get_permissions}} to check GitHub permissions
}

