% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/account.R
\name{get_users}
\alias{get_users}
\alias{sync_users}
\title{Travis CI Users}
\usage{
get_users(user, ...)

sync_users(...)
}
\arguments{
\item{user}{Optionally, a character string specifying a user. If missing, all users are returned.}

\item{...}{Additional arguments passed to \code{\link{travisHTTP}}.}
}
\value{
A list.
}
\description{
Retrieve and sync Travis CI users
}
\details{
\code{get_users} retrieves information about GitHub users attached to a Travis account. \code{sync_users} syncs Travis's local cache of users against GitHub.
}
\examples{
\dontrun{
# authenticate based on Sys.setenv("GITHUB_TOKEN" = "sometoken")
auth_travis()

# sync users
sync_users()

# get travis users
get_users()

# get specific user
get_user("userid")
}
}
\seealso{
\code{\link{get_accounts}}, \code{\link{get_permissions}}
}

