% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{travis_lint}
\alias{travis_lint}
\title{Linting}
\usage{
travis_lint(file, ...)
}
\arguments{
\item{file}{A character string specifying a path to a \samp{.travis.yml} file.}

\item{...}{Additional arguments passed to \code{\link{travisHTTP}}.}
}
\value{
A list.
}
\description{
Lint a .travis.yml file
}
\details{
This can validate a \samp{.travis.yml} file and identify possible errors.
}
\examples{
\dontrun{
# authenticate based on Sys.setenv("GITHUB_TOKEN" = "sometoken")
auth_travis()

# validate .travis.yml
travis_lint("./.travis.yml")
}
}
\seealso{
\code{\link{get_permissions}} to check GitHub permissions
}

