% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{travisHTTP}
\alias{travisHTTP}
\title{Travis-CI HTTP Requests}
\usage{
travisHTTP(verb = "GET", path = "", query = list(), body = "",
  base = c("https://api.travis-ci.org", "https://api.travis-ci.com"),
  token = Sys.getenv("TRAVIS_CI_TOKEN"), ...)
}
\arguments{
\item{verb}{A character string containing an HTTP verb, defaulting to \dQuote{GET}.}

\item{path}{A character string with the API endpoint (should begin with a slash).}

\item{query}{A list specifying any query string arguments to pass to the API.}

\item{body}{A character string of request body data.}

\item{base}{A character string specifying the base URL for the API.}

\item{token}{A character string containing a Travis-CI API token. If missing, defaults to value stored in environment variable \env{TRAVIS_CI_TOKEN}.}

\item{...}{Additional arguments passed to an HTTP request function, such as \code{\link[httr]{GET}}.}
}
\value{
A list.
}
\description{
This is the workhorse function for executing API requests for Travis-CI.
}
\details{
This is mostly an internal function for executing API requests. In almost all cases, users do not need to access this directly.
}

