% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/builds.R
\name{cancel_build}
\alias{cancel_build}
\alias{restart_build}
\title{Cancel and Restart Builds}
\usage{
cancel_build(build, ...)

restart_build(build, ...)
}
\arguments{
\item{build}{A numeric value specifying a build number.}

\item{...}{Additional arguments passed to \code{\link{travisHTTP}}.}
}
\value{
A list.
}
\description{
Cancel and restart Travis-CI builds
}
\details{
\code{cancel_build} will cancel a given build. \code{restart_build} will restart a cancelled build.
}
\examples{
\dontrun{
# get all recent builds for this repo
g <- get_builds(repo = "cloudyr/travisci")

# cancel most recent build
cancel_build(g[[1]]$id)

# restart that build
restart_build(g[[1]]$id)
}
}
\seealso{
\code{\link{get_builds}}
}

