% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repo.R
\name{get_branch}
\alias{get_branch}
\title{Get Branch}
\usage{
get_branch(repo, branch, ...)
}
\arguments{
\item{repo}{Optionally, a numeric repository ID (such as returned by this function) or a character string specifying a GitHub repository \dQuote{slug} (e.g., \samp{ghusername/ghreponame}).}

\item{branch}{A character string specifying the name of a branch.}

\item{...}{Additional arguments passed to \code{\link{travisHTTP}}.}
}
\value{
A list.
}
\description{
Retrieve branches for a repo
}
\details{
This can retrieve a list of recent branches for a given repo, or if \code{branch} is specified, details about a specific branch.
}
\examples{
\dontrun{
# get branches for a given repo
get_branch(repo = "cloudyr/travisci")

# get a specific branch
get_branch(repo = "cloudyr/travisci", "master")
}
}
\seealso{
\code{\link{get_builds}}

\code{\link{get_repo}}, \code{\link{get_builds}}
}

