% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/environment.R
\name{get_env_vars}
\alias{add_env_vars}
\alias{delete_env_vars}
\alias{get_env_vars}
\alias{update_env_vars}
\title{Manage environment variables}
\usage{
get_env_vars(repo, id, ...)

add_env_vars(repo, var = list(), ...)

update_env_vars(repo, id, var = list(), ...)

delete_env_vars(repo, id, ...)
}
\arguments{
\item{repo}{A character string specifying a repo slug (i.e., \samp{cloudyr/travisci}) or a numeric Travis-CI repository ID. If a slug is used, it will be implicitly converted to a repository ID in some cases where only the latter is accepted.}

\item{id}{A numeric ID for a given environment variable, for example as returned by this function with this argument missing.}

\item{...}{Additional arguments passed to \code{\link{travisHTTP}}.}

\item{var}{A list of environment variables as key-value pairs.}
}
\value{
A list.
}
\description{
These functions retrieve and modify environment variables for the Travis-CI build environment, as a possible alternative to specifying them in a \samp{.travis.yml} file.
}
\details{
This can be useful for checking, creating, updating, or deleting environment variables used as build settings. See href{http://docs.travis-ci.com/user/environment-variables/}{the API documentation} for full details.
}
\seealso{
\code{\link{get_encryption_key}}
}

