% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/requests.R
\name{get_requests}
\alias{get_requests}
\title{GitHub Requests}
\usage{
get_requests(request, ...)
}
\arguments{
\item{request}{Optionally, a numeric request ID.}

\item{...}{Additional arguments passed to \code{\link{travisHTTP}}.}
}
\value{
A list.
}
\description{
Examine requests from GitHub to Travis CI
}
\details{
This can examine requests made by GitHub to initiate builds on Travis-CI, perhaps when a push fails to execute a build.
}
\examples{
\dontrun{
# authenticate based on Sys.setenv("GITHUB_TOKEN" = "sometoken")
auth_travis()

# get a list of requests
get_requests()

# get a specific request
get_requests(12345)
}
}
\seealso{
\code{\link{get_permissions}} to check GitHub permissions
}

